+++ analyze_test.go
@@ -16,6 +16,7 @@ func TestBuildArgs(t *testing.T) {
 		name          string
 		excludedPaths string
 		enableMetrics bool
+		passedCLIOpts string
 		want          []string
 	}{
 		{
@@ -113,13 +114,97 @@ func TestBuildArgs(t *testing.T) {
 				"--enable-metrics",
 			},
 		},
+		{
+			name:          "CLI Opts: append value to flag name with = separator",
+			excludedPaths: "",
+			passedCLIOpts: "--max-memory=1024",
+			enableMetrics: true,
+			want: []string{
+				"-f", "configPath",
+				"-o", "outputPath",
+				"--sarif",
+				"--no-rewrite-rule-ids",
+				"--strict",
+				"--disable-version-check",
+				"--no-git-ignore",
+				"--enable-metrics",
+				"--max-memory", "1024",
+			},
+		},
+		{
+			name:          "CLI Opts: Include both allowed and not-allowed flag with = separator",
+			excludedPaths: "",
+			passedCLIOpts: "--max-memory=1024 --optimizations=all", // invalid
+			enableMetrics: true,
+			want: []string{
+				"-f", "configPath",
+				"-o", "outputPath",
+				"--sarif",
+				"--no-rewrite-rule-ids",
+				"--strict",
+				"--disable-version-check",
+				"--no-git-ignore",
+				"--enable-metrics",
+				"--max-memory", "1024",
+			},
+		},
+		{
+			name:          "CLI Opts: Include only allowed flag with space separator",
+			excludedPaths: "",
+			passedCLIOpts: "--max-memory 1024",
+			enableMetrics: true,
+			want: []string{
+				"-f", "configPath",
+				"-o", "outputPath",
+				"--sarif",
+				"--no-rewrite-rule-ids",
+				"--strict",
+				"--disable-version-check",
+				"--no-git-ignore",
+				"--enable-metrics",
+				"--max-memory", "1024",
+			},
+		},
+		{
+			name:          "CLI Opts: Include both allowed and not-allowed flag with space separator",
+			excludedPaths: "",
+			passedCLIOpts: "--max-memory 1024 --optimizations all",
+			enableMetrics: true,
+			want: []string{
+				"-f", "configPath",
+				"-o", "outputPath",
+				"--sarif",
+				"--no-rewrite-rule-ids",
+				"--strict",
+				"--disable-version-check",
+				"--no-git-ignore",
+				"--enable-metrics",
+				"--max-memory", "1024",
+			},
+		},
+		{
+			name:          "CLI Opts: Include only not-allowed flag",
+			excludedPaths: "",
+			passedCLIOpts: "--optimizations=all",
+			enableMetrics: true,
+			want: []string{
+				"-f", "configPath",
+				"-o", "outputPath",
+				"--sarif",
+				"--no-rewrite-rule-ids",
+				"--strict",
+				"--disable-version-check",
+				"--no-git-ignore",
+				"--enable-metrics",
+			},
+		},
 	}
 
 	for _, tt := range tests {
 		tt := tt // NOTE: https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			got := buildArgs("configPath", "outputPath", "projectPath", tt.excludedPaths, tt.enableMetrics)
+			got := buildArgs("configPath", "outputPath", "projectPath", tt.excludedPaths, tt.passedCLIOpts, tt.enableMetrics)
 
 			if !reflect.DeepEqual(tt.want, got) {
 				t.Errorf("Wrong result. Expected:\n%#v\nbut got:\n%#v", tt.want, got)